Implemented LoadEmojiFile func and updated data.
Implemented a LoadEmojiFile func that allows the developer to ship and load the emoji.json file in a separate directory than the vendor expected path of the executable / caller + data/emoji.json (or data\emoji.json on Windows). Utilize this func in our Init as well to unify expected behavior, retaining the expected panic if we successfully unmarshal the provided file in LoadEmojiFile (including during init) however with the result of an empty Emojis map.